Moved the create data frame code into a function. Added argument options to the regressor command.
