+++ tcmg/Summary.swift
@@ -18,29 +18,11 @@ extension TCMG {
         @Flag var columnNames: Bool = false
         
         public func run() throws {
-            
-            let csvOptions = CSVReadingOptions(
-                hasHeaderRow: true,
-                delimiter: ","
-            )
-            
-            var summaryDataFrame: DataFrame = DataFrame()
-            
-            do {
-                summaryDataFrame = try DataFrame(
-                    contentsOfCSVFile: newCSVFileURL(options.dataFileName),
-                    options: csvOptions
-                )
-            } catch UtilityError.fileWrongType {
-                print("ERROR: NOT A VALID CSV FILE")
-            } catch {
-                print("ERROR: AN UNKNOWN ERROR")
-            }
-            
-            print(columnNames ? summaryDataFrame.columns.map { colName in colName.name } : summaryDataFrame.summary())
-            
-            //print(summaryDataFrame.summary())
-            //print(summaryDataFrame.columns.map { colName in colName.name })
+            var summaryDataFrame = try createCSVDataFrame(options.dataFileName)
+             
+            print(columnNames ? summaryDataFrame.columns.map { colName in
+                colName.name
+            } : summaryDataFrame.summary())
         }
     }
 }
